requirements.txt
asgiref==3.9.1
Django==4.2.23
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.0
importlib_metadata==8.7.0
Markdown==3.8.2
psycopg2-binary==2.9.10
sqlparse==0.5.3
typing_extensions==4.14.1
tzdata==2025.2
zipp==3.23.0
